pb-notify: implement clipboard integration
pb-notify now maintains an internal buffer to represent the pushbullet
clipboard. When Clipboard events come down the stream, this buffer is
updated.
An HTTP server is added to pb-notify.
A single endpoint is implemented, allowed a user to retrieve the
contents of the buffer and to broadcast a new clipboard synchronization
message to PushBullet. Other devices configured to use Universal
Copy/Paste will receive this text in their clipboards.

Note that no interaction with the system clipboard is currently
provided.